+++ src/config/index.ts
@@ -18,12 +18,19 @@ export { default } from '@v4fire/core/config';
 export * from '@v4fire/core/config';
 export * from 'config/interface';
 
+const IS_SSR = Boolean(typeof SSR !== 'undefined' && SSR);
+
 extend({
 	image: {},
 
 	asyncRender: {
 		weightPerTick: 5,
 		delay: 40
+	},
+
+	gc: {
+		quota: IS_SSR ? 50 : 25,
+		delay: IS_SSR ? 1000 : 5000
 	},
 
 	/**