+++ src/core/component/engines/vue3/component.ts
@@ -13,7 +13,7 @@ import * as init from 'core/component/init';
 import { beforeRenderHooks } from 'core/component/const';
 
 import { fillMeta } from 'core/component/meta';
-import { getComponentContext } from 'core/component/context';
+import { getComponentContext, dropRawComponentContext } from 'core/component/context';
 import { wrapAPI } from 'core/component/render';
 
 import type { ComponentEngine, ComponentOptions, SetupContext } from 'core/component/engines';
@@ -83,12 +83,13 @@ export function getComponent(meta: ComponentMeta): ComponentOptions<typeof Compo
 			const
 				ctx = getComponentContext(this);
 
-			Object.set(ctx, '$renderEngine', {
+			// @ts-ignore (unsafe)
+			ctx['$renderEngine'] = {
 				supports,
 				proxyGetters,
 				r,
 				wrapAPI
-			});
+			};
 
 			init.beforeCreateState(ctx, meta, {implementEventAPI: true});
 
@@ -141,11 +142,26 @@ export function getComponent(meta: ComponentMeta): ComponentOptions<typeof Compo
 		},
 
 		beforeUnmount(): void {
-			init.beforeDestroyState(getComponentContext(this), {recursive: false});
+			const ctx = getComponentContext(this);
+
+			// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+			if (ctx == null) {
+				return;
+			}
+
+			init.beforeDestroyState(ctx, {recursive: false});
 		},
 
 		unmounted(): void {
-			init.destroyedState(getComponentContext(this));
+			const ctx = getComponentContext(this);
+
+			// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+			if (ctx == null) {
+				return;
+			}
+
+			init.destroyedState(ctx);
+			dropRawComponentContext(ctx);
 		},
 
 		errorCaptured(...args: unknown[]): void {