+++ src/core/component/init/states/before-data-create.ts
@@ -21,14 +21,12 @@ export function beforeDataCreateState(component: ComponentInterface): void {
 	const {meta, $fields} = component.unsafe;
 	initFields(meta.fields, component, $fields);
 
-	// Because in functional components,
-	// the watching of fields can be initialized in lazy mode
+	// In functional components, the watching of fields can be initialized in lazy mode
 	if (SSR || meta.params.functional === true) {
 		Object.assign(component, $fields);
 	}
 
-	runHook('beforeDataCreate', component)
-		.catch(stderr);
+	runHook('beforeDataCreate', component).catch(stderr);
 
 	if (!SSR) {
 		implementComponentWatchAPI(component);