+++ src/core/component/functional/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.121.the-phantom-menace (2024-08-05)
+
+#### :bug: Bug Fix
+
+* Fixed an issue with updating modifier values
+
 ## v4.0.0-beta.22 (2023-09-15)
 
 #### :bug: Bug Fix